ref: move callbackify to lib/util.ts file
